organized src into packages so its easier to navigate
