+++ server/src/dao/model/incoming_image.py
